fix compiler warnings extend API minor changes

+++ libs/qcrypt/src/qcrypt.c
@@ -292,7 +292,7 @@ CapeString qcrypt__encode_base64_o (const char* bufdat, number_t buflen)
   CapeString ret = CAPE_ALLOC (len);
   
   // openssl function
-  int decodedSize = EVP_EncodeBlock ((unsigned char*)ret, (const unsigned char*)bufdat, buflen);
+  int decodedSize = EVP_EncodeBlock ((unsigned char*)ret, (const unsigned char*)bufdat, (int)buflen);
   
   // everything worked fine
   if ((decodedSize > 0) && (decodedSize < len))
@@ -330,7 +330,7 @@ CapeStream qcrypt__decode_base64_o (const char* bufdat, number_t buflen)
   cape_stream_cap (ret, len);
   
   // openssl function
-  int decodedSize = EVP_DecodeBlock ((unsigned char*)cape_stream_pos (ret), (const unsigned char*)bufdat, buflen);
+  int decodedSize = EVP_DecodeBlock ((unsigned char*)cape_stream_pos (ret), (const unsigned char*)bufdat, (int)buflen);
   
   // everything worked fine
   if ((decodedSize > 0) && (decodedSize < len))
